% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FiguresAndTables.R
\name{createMetaData}
\alias{createMetaData}
\title{Create metadata file}
\usage{
createMetaData(addCdmSource = TRUE, connectionDetails, cdmDatabaseSchema,
  addPackageVersions = TRUE, package, exportFolder)
}
\arguments{
\item{addCdmSource}{Add information from the CDM_SOURCE table?}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{addPackageVersions}{Add information on the versions of dependency packages?}

\item{package}{Name of the package for which we want to list the versions of its dependencies.#'}

\item{exportFolder}{The name of the folder where the metadata file should be created.}
}
\description{
Create metadata file
}
\details{
Creates a file containing metadata about the source data (taken from the cdm_source table) and R
package versions.
}
